UI is way more functional
server will now properly update state on a victory.
it does not know anything about draws atm though.

Also resignations are not accepted.

+++ service/pyShobuService/src/logic/shobu/helpers.py
@@ -4,7 +4,7 @@ import logic.shobu.token as t
 def vectorToUnitAndSteps( vector ):
     x,y = vector
     if max( abs( x ), abs( y ) ) == 2:
-        return (x/2,y/2),2
+        return (int(x/2),int(y/2)),2
     return (x,y),1
 
 def tokenIsEnemy( token, player ):
